Move nullif to arrow-select (#2594)
